pyproject.toml
[tool.poetry]
name = "acdc"
version = "0.1.1"
description = "Wanna-be fast string search"
authors = ["Amirouche <amirouche@hyper.dev>"]
license = "MIT"
readme = "README.md"
repository = "https://git.sr.ht/~amirouche/python-acdc"

[tool.poetry.dependencies]
python = "^3.7"

[tool.poetry.dev-dependencies]
memory-profiler = "^0.58.0"
ahocorapy = "^1.6.1"
scalene = "^1.3.8"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
debug = "^0.3.2"
awscli = "^1.19.105"
html2text = "^2020.1.16"
ipython = "^7.25.0"
termgraph = "^0.5.1"
python-slugify = "^5.0.2"
pyahocorasick = "^1.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
